+++ src/test/testConfiguration.js
@@ -20,7 +20,7 @@ exports.configuration = new __1.Configuration({
     database: {
         MongoDB: {
             db_url: envirn.MONGO_DB_URL,
-            db: 'bpmn'
+            db: 'bpmnTest'
         }
     },
     apiKey: envirn.API_KEY,