+++ src/views/InstanceViewer.ts
@@ -73,7 +73,10 @@ function getHeader() {
 	<script type='text/javascript' src='/vendor/dhtmlx/codebase/dhtmlx.js'></script>
 	<script type='text/javascript' src='/vendor/dhtmlx/codebase/datastore.js'></script>
 	<script type='text/javascript' src='/javascripts/jsonHelper.js'></script>
+	 <!-- GSAP Library from CDN -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
 	<script type='text/javascript' src='/javascripts/SVGHelper.js'></script>
+
 	<link rel='stylesheet' href='/vendor/dhtmlx/codebase/dhtmlx.css' type='text/css'>
 	<link rel='stylesheet' href='/vendor/dhtmlx/skins/skyblue/dhtmlx.css' type='text/css'>
 </head>